move pressure control + waste detector into separate classes to make them easier to midify and test

+++ stm32-modules/include/vacuum-module/firmware/slew_rate_limiter.hpp
@@ -3,7 +3,7 @@
 
 class SlewRateLimiter {
   public:
-    SlewRateLimiter(float current = 0.0F, float rate_limit = 0.0F)
+    SlewRateLimiter(double current = 0.0F, double rate_limit = 0.0F)
         : _current_output(current), _rate_limit(rate_limit) {}
 
     /**
@@ -12,7 +12,7 @@ class SlewRateLimiter {
      * * @param start_value The starting point (e.g. current actual pressure).
      * @param rate_limit_per_sec The max change allowed per second.
      */
-    auto configure(float start_value, float rate_limit_per_sec) -> void {
+    auto configure(double start_value, double rate_limit_per_sec) -> void {
         _current_output = start_value;
         _rate_limit = rate_limit_per_sec;
     }
@@ -22,13 +22,13 @@ class SlewRateLimiter {
      * Call this when the control loop resumes from Idle to prevent jumps.
      * Does not change the configured rate limit.
      */
-    auto reset(float value = 0.0) -> void { _current_output = value; }
+    auto reset(double value = 0.0) -> void { _current_output = value; }
 
     /**
      * @brief Update the rate limit dynamically.
      * Useful if you have a "Fast Mode" and "Slow Mode".
      */
-    auto set_rate_limit(float rate_limit_per_sec) -> void {
+    auto set_rate_limit(double rate_limit_per_sec) -> void {
         _rate_limit = rate_limit_per_sec;
     }
 
@@ -38,7 +38,7 @@ class SlewRateLimiter {
      * @param dt_seconds Time elapsed since last call.
      * @return The smoothed output.
      */
-    auto update(float target, float dt_seconds) -> float {
+    auto update(double target, double dt_seconds) -> double {
         auto const error = target - _current_output;
         auto const max_change = _rate_limit * dt_seconds;
 
@@ -53,11 +53,11 @@ class SlewRateLimiter {
         return _current_output;
     }
 
-    [[nodiscard]] auto get_current_setpoint() const -> float {
+    [[nodiscard]] auto get_current_setpoint() const -> double {
         return _current_output;
     }
 
   private:
-    float _current_output = 0.0;
-    float _rate_limit = 0.0;
+    double _current_output = 0.0;
+    double _rate_limit = 0.0;
 };
